implement add new subcategory to a category and category edit

+++ Services/AYN.Services.Data/CategoriesService.cs
@@ -17,14 +17,17 @@ namespace AYN.Services.Data
     {
         private readonly IDeletableEntityRepository<Category> categoriesRepository;
         private readonly IDeletableEntityRepository<SubCategory> subCategoriesRepository;
+        private readonly ISubCategoriesService subCategoriesService;
         private readonly string[] allowedImageExtensions = { "jpg", "png", "jfif", "exif", "gif", "bmp", "ppm", "pgm", "pbm", "pnm", "heif", "bat" };
 
         public CategoriesService(
             IDeletableEntityRepository<Category> categoriesRepository,
-            IDeletableEntityRepository<SubCategory> subCategoriesRepository)
+            IDeletableEntityRepository<SubCategory> subCategoriesRepository,
+            ISubCategoriesService subCategoriesService)
         {
             this.categoriesRepository = categoriesRepository;
             this.subCategoriesRepository = subCategoriesRepository;
+            this.subCategoriesService = subCategoriesService;
         }
 
         public async Task CreateAsync(CreateCategoryInputModel input, string imagePath)
@@ -62,6 +65,11 @@ namespace AYN.Services.Data
                         Name = subCategory.Name,
                     };
 
+                    if (this.subCategoriesService.IsSubCategoryExisting(currentSubCategory.Name))
+                    {
+                        throw new InvalidOperationException($"This subCategory '{currentSubCategory.Name}' is already taken by another category!");
+                    }
+
                     category.SubCategories
                         .Add(currentSubCategory);
                 }
@@ -85,11 +93,38 @@ namespace AYN.Services.Data
                 .All()
                 .To<T>();
 
+        public async Task AddSubCategoryAsync(AddSubCategoryViewModel input, int categoryId)
+            => await this.subCategoriesService.CreateAsync(input, categoryId);
+
+        public T GetById<T>(int categoryId)
+            => this.categoriesRepository
+                .All()
+                .Where(c => c.Id == categoryId)
+                .To<T>()
+                .FirstOrDefault();
+
+        public async Task UpdateAsync(EditCategoryInputModel input, int categoryId)
+        {
+            var category = this.categoriesRepository
+                .All()
+                .FirstOrDefault(c => c.Id == categoryId);
+
+            category.Name = input.Name;
+
+            this.categoriesRepository.Update(category);
+            await this.categoriesRepository.SaveChangesAsync();
+        }
+
         public async Task DeleteAsync(int id)
         {
             var category = this.categoriesRepository
                 .All()
                 .FirstOrDefault(c => c.Id == id);
+
+            if (category == null)
+            {
+                throw new InvalidOperationException($"The category doesn't exist!");
+            }
 
             this.categoriesRepository.Delete(category);
